Cluster policy tweaks.
