+++ src/Thycotic.RabbitMq.Helper.PSCommands/Thycotic.RabbitMq.Helper.PSCommands.csproj
@@ -216,7 +216,7 @@
     <Compile Include="Certificate\ConvertCaCerToPemCommand.cs" />
     <Compile Include="Certificate\ConvertPfxToPemCommand.cs" />
     <Compile Include="Clustering\JoinRabbitMqClusterCommand.cs" />
-    <Compile Include="Clustering\Policy\SetRabbitMqBalancedOneMirrorManualSyncClusterPolicyCommand.cs" />
+    <Compile Include="Clustering\Policy\SetRabbitMqBalancedClusterPolicyCommand.cs" />
     <Compile Include="Clustering\SetErlangCookieFileCommand.cs" />
     <Compile Include="Clustering\GetRabbitMqClusterNodesCommand.cs" />
     <Compile Include="CommandAreas.cs" />
